latest version - waiting for tree evaluation features to be put in by zach
also please note - if you download this and work on it locally, you must relink the solutions. the solutions are currently linked with a path that involves Nathan De Jong in the path name b/c it was on my computer - you will get linker errors if you don't do this

right click on each main item (DBMS, Main, TableTester), go to properties -> configuration properties -> vc++ directories. in the include directory line, remove what is there, and then include a new directory which is will end as "../DBMS/DBMS" - this should work and fix the linker errors